# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description of my API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/create_agent": {
            "post": {
                "tags": [
                    "Admin Agents"
                ],
                "summary": "Create a new agent",
                "description": "Creates a new agent with the specified email, name, surname, and phone.",
                "operationId": "9923ccc68e09916685c3fc8d21e91d18",
                "requestBody": {
                    "description": "Agent details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Agent's email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "name": {
                                        "description": "Agent's name",
                                        "type": "string"
                                    },
                                    "surname": {
                                        "description": "Agent's surname",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Agent's phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Agent created successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/update_agent": {
            "post": {
                "tags": [
                    "Admin Agents"
                ],
                "summary": "Update agent information",
                "description": "Updates the information of an existing agent based on the provided agent_id.",
                "operationId": "118f1dd09b0bd32dd58f80f492ec6fd0",
                "requestBody": {
                    "description": "Agent details to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "New email for the agent",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "name": {
                                        "description": "New name for the agent",
                                        "type": "string"
                                    },
                                    "surname": {
                                        "description": "New surname for the agent",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "New phone number for the agent",
                                        "type": "string"
                                    },
                                    "agent_id": {
                                        "description": "ID of the agent to update",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent information updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Agent updated successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/single_page_agent": {
            "get": {
                "tags": [
                    "Admin Agents"
                ],
                "summary": "Get information of a single agent",
                "description": "Retrieves information of a single agent based on the provided agent_id.",
                "operationId": "dd26ccd3dfde84a5fb0e762ad2f374ac",
                "parameters": [
                    {
                        "name": "agent_id",
                        "in": "query",
                        "description": "ID of the agent to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agent not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Agent Not Found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/all_agents": {
            "get": {
                "tags": [
                    "Admin Agents"
                ],
                "summary": "Get all agents with optional search",
                "description": "Retrieves a paginated list of all agents with optional search functionality.",
                "operationId": "bdcd5ee0babaee2a398fd36e91b61ee7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keyword to filter agents by name, surname, phone, or email",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    }
}